moves input components to their own folder and deletes old stylesheet
