--- src/home.css
@@ -1,72 +1,0 @@
-html,
-body {
-  height: 100%;
-  margin: 0;
-  padding: 0;
-}
-
-.navbar {
-  border-radius: 8px;
-}
-
-#map {
-  height: 400px;
-}
-
-#compute-div {
-  text-align: center;
-  display: block;
-}
-
-#compute-btn, .navbar {
-  background-color: #53afed;
-}
-
-.input-group, .mb-3 {
-  margin-top: 10px;
-  margin-bottom: 10px;
-}
-
-#info {
-  margin-bottom: 20px;
-}
-
-@media (max-width: 575.98px) {
-  .col-5, .col-7 {
-    max-width: 100%;
-  }
-}
-
-#github-logo {
-  max-height: 40px;
-}
-
-#flex {
-  padding: 0;
-}
-
-/* Small devices (landscape phones, 576px and up) */
-@media (min-width: 576px) {}
-
-/* Medium devices (tablets, 768px and up) */
-@media (min-width: 768px) {}
-
-/* Large devices (desktops, 992px and up) */
-@media (min-width: 992px) {
-  .btn {
-    margin-top: 10px;
-    margin-bottom: 10px;
-  }
-
-  #flex {
-    display: flex;
-    padding: 0;
-  }
-
-  #map {
-    height: 500px;
-  }
-}
-
-/* Extra large devices (large desktops, 1200px and up) */
-@media (min-width: 1200px) {}
